Add flag for WAL mode, db.SetMaxOpenConns(1) and use real files
This change drops the use of memory mode in favor of real files on disk.
These files are cleaned up after the run. Also taking tips from the
go-sqlite README I added:

- db.SetMaxOpenConns(1)
- ?cached=shared

Using the db.SetMaxOpenConns(1) with no locking (-type=none) and WAL mode
(-wal) resulted in the fastest tests.  This seems to be the preferred
way.  In real app code the sync.Mutex are not required.  Just open the
database in WAL mode with SetMaxOpenConns(1)